+++ internal/cmd/company/iam/group/remove_member.go
@@ -26,14 +26,14 @@ import (
 )
 
 const (
-	addMemberTemplate = "/api/companies/%s/groups/%s/members"
+	removeMemberTemplate = "/api/companies/%s/groups/%s/members"
 )
 
-func AddMemberCmd(options *clioptions.CLIOptions) *cobra.Command {
+func RemoveMemberCmd(options *clioptions.CLIOptions) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "group-member",
-		Short: "Add one or more users to a group",
-		Long:  "Add one or more users to a company group. The users can be added via their emails",
+		Short: "Remove one or more users from a group",
+		Long:  "Remove one or more users from a company group. The users can be removed via their ids",
 
 		Args: cobra.NoArgs,
 		Run: func(cmd *cobra.Command, args []string) {
@@ -42,16 +42,16 @@ func AddMemberCmd(options *clioptions.CLIOptions) *cobra.Command {
 			client, err := client.APIClientForConfig(restConfig)
 			cobra.CheckErr(err)
 
-			err = addMemberToGroup(cmd.Context(), client, restConfig.CompanyID, options.GroupID, options.UserEmails)
+			err = removeMemberFromGroup(cmd.Context(), client, restConfig.CompanyID, options.GroupID, options.UserIDs)
 			cobra.CheckErr(err)
 		},
 	}
 
-	options.AddNewMembersToGroupFlags(cmd.Flags())
+	options.AddRemoveMembersFromGroupFlags(cmd.Flags())
 	return cmd
 }
 
-func addMemberToGroup(ctx context.Context, client *client.APIClient, companyID, groupID string, userEmails []string) error {
+func removeMemberFromGroup(ctx context.Context, client *client.APIClient, companyID, groupID string, userIDs []string) error {
 	if len(companyID) == 0 {
 		return fmt.Errorf("company id is required, please set it via flag or context")
 	}
@@ -60,12 +60,12 @@ func addMemberToGroup(ctx context.Context, client *client.APIClient, companyID, 
 		return fmt.Errorf("a group id is required")
 	}
 
-	if len(userEmails) < 1 {
-		return fmt.Errorf("at least one user must be added to the group")
+	if len(userIDs) < 1 {
+		return fmt.Errorf("at least one user id must be used")
 	}
 
-	payload := resources.AddMembersToGroup{
-		Members: userEmails,
+	payload := resources.RemoveMembersToGroup{
+		Members: userIDs,
 	}
 
 	body, err := resources.EncodeResourceToJSON(payload)
@@ -74,8 +74,8 @@ func addMemberToGroup(ctx context.Context, client *client.APIClient, companyID, 
 	}
 
 	resp, err := client.
-		Post().
-		APIPath(fmt.Sprintf(addMemberTemplate, companyID, groupID)).
+		Delete().
+		APIPath(fmt.Sprintf(removeMemberTemplate, companyID, groupID)).
 		Body(body).
 		Do(ctx)
 
@@ -87,6 +87,6 @@ func addMemberToGroup(ctx context.Context, client *client.APIClient, companyID, 
 		return err
 	}
 
-	fmt.Println("the users has been added to the group")
+	fmt.Println("the users has been removed from the group")
 	return nil
 }