+++ internal/cmd/company/iam/serviceaccount/remove.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package group
+package serviceaccount
 
 import (
 	"context"
@@ -21,19 +21,18 @@ import (
 
 	"github.com/mia-platform/miactl/internal/client"
 	"github.com/mia-platform/miactl/internal/clioptions"
-	"github.com/mia-platform/miactl/internal/resources"
 	"github.com/spf13/cobra"
 )
 
 const (
-	addMemberTemplate = "/api/companies/%s/groups/%s/members"
+	removeServiceAccountTemplate = "/api/companies/%s/service-accounts/%s"
 )
 
-func AddMemberCmd(options *clioptions.CLIOptions) *cobra.Command {
+func RemoveCmd(options *clioptions.CLIOptions) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "group-member",
-		Short: "Add one or more users to a group",
-		Long:  "Add one or more users to a company group. The users can be added via their emails",
+		Use:   "serviceaccount",
+		Short: "Remove a service account from a company",
+		Long:  "Remove a service account from a company",
 
 		Args: cobra.NoArgs,
 		Run: func(cmd *cobra.Command, args []string) {
@@ -42,42 +41,29 @@ func AddMemberCmd(options *clioptions.CLIOptions) *cobra.Command {
 			client, err := client.APIClientForConfig(restConfig)
 			cobra.CheckErr(err)
 
-			err = addMemberToGroup(cmd.Context(), client, restConfig.CompanyID, options.GroupID, options.UserEmails)
+			err = removeCompanyServiceAccount(cmd.Context(), client, restConfig.CompanyID, options.ServiceAccountID)
 			cobra.CheckErr(err)
 		},
 	}
 
-	options.AddNewMembersToGroupFlags(cmd.Flags())
+	options.AddRemoveServiceAccountFlags(cmd.Flags())
 	return cmd
 }
 
-func addMemberToGroup(ctx context.Context, client *client.APIClient, companyID, groupID string, userEmails []string) error {
+func removeCompanyServiceAccount(ctx context.Context, client *client.APIClient, companyID, serviceAccountID string) error {
 	if len(companyID) == 0 {
 		return fmt.Errorf("company id is required, please set it via flag or context")
 	}
 
-	if len(groupID) == 0 {
-		return fmt.Errorf("a group id is required")
+	if len(serviceAccountID) == 0 {
+		return fmt.Errorf("the service account id is required")
 	}
 
-	if len(userEmails) < 1 {
-		return fmt.Errorf("at least one user must be added to the group")
-	}
+	request := client.
+		Delete().
+		APIPath(fmt.Sprintf(removeServiceAccountTemplate, companyID, serviceAccountID))
 
-	payload := resources.AddMembersToGroup{
-		Members: userEmails,
-	}
-
-	body, err := resources.EncodeResourceToJSON(payload)
-	if err != nil {
-		return fmt.Errorf("failed to encode request body: %w", err)
-	}
-
-	resp, err := client.
-		Post().
-		APIPath(fmt.Sprintf(addMemberTemplate, companyID, groupID)).
-		Body(body).
-		Do(ctx)
+	resp, err := request.Do(ctx)
 
 	if err != nil {
 		return err
@@ -87,6 +73,6 @@ func addMemberToGroup(ctx context.Context, client *client.APIClient, companyID, 
 		return err
 	}
 
-	fmt.Println("the users has been added to the group")
+	fmt.Printf("service account %s successfully removed\n", serviceAccountID)
 	return nil
 }