+++ internal/cmd/company/iam/user/remove.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package group
+package user
 
 import (
 	"context"
@@ -21,19 +21,19 @@ import (
 
 	"github.com/mia-platform/miactl/internal/client"
 	"github.com/mia-platform/miactl/internal/clioptions"
-	"github.com/mia-platform/miactl/internal/resources"
 	"github.com/spf13/cobra"
 )
 
 const (
-	addMemberTemplate = "/api/companies/%s/groups/%s/members"
+	removeUserTemplate      = "/api/companies/%s/users/%s"
+	removeFromGroupParamKey = "removeFromGroups"
 )
 
-func AddMemberCmd(options *clioptions.CLIOptions) *cobra.Command {
+func RemoveCmd(options *clioptions.CLIOptions) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "group-member",
-		Short: "Add one or more users to a group",
-		Long:  "Add one or more users to a company group. The users can be added via their emails",
+		Use:   "user",
+		Short: "Remove a user from a company",
+		Long:  "Remove a user from a company",
 
 		Args: cobra.NoArgs,
 		Run: func(cmd *cobra.Command, args []string) {
@@ -42,42 +42,33 @@ func AddMemberCmd(options *clioptions.CLIOptions) *cobra.Command {
 			client, err := client.APIClientForConfig(restConfig)
 			cobra.CheckErr(err)
 
-			err = addMemberToGroup(cmd.Context(), client, restConfig.CompanyID, options.GroupID, options.UserEmails)
+			err = removeCompanyUser(cmd.Context(), client, restConfig.CompanyID, options.UserID, options.KeepUserGroupMemeberships)
 			cobra.CheckErr(err)
 		},
 	}
 
-	options.AddNewMembersToGroupFlags(cmd.Flags())
+	options.AddRemoveUserFlags(cmd.Flags())
 	return cmd
 }
 
-func addMemberToGroup(ctx context.Context, client *client.APIClient, companyID, groupID string, userEmails []string) error {
+func removeCompanyUser(ctx context.Context, client *client.APIClient, companyID, userID string, keepMemberships bool) error {
 	if len(companyID) == 0 {
 		return fmt.Errorf("company id is required, please set it via flag or context")
 	}
 
-	if len(groupID) == 0 {
-		return fmt.Errorf("a group id is required")
+	if len(userID) == 0 {
+		return fmt.Errorf("the user id is required")
 	}
 
-	if len(userEmails) < 1 {
-		return fmt.Errorf("at least one user must be added to the group")
+	request := client.
+		Delete().
+		APIPath(fmt.Sprintf(removeUserTemplate, companyID, userID))
+
+	if !keepMemberships {
+		request.SetParam(removeFromGroupParamKey, "true")
 	}
 
-	payload := resources.AddMembersToGroup{
-		Members: userEmails,
-	}
-
-	body, err := resources.EncodeResourceToJSON(payload)
-	if err != nil {
-		return fmt.Errorf("failed to encode request body: %w", err)
-	}
-
-	resp, err := client.
-		Post().
-		APIPath(fmt.Sprintf(addMemberTemplate, companyID, groupID)).
-		Body(body).
-		Do(ctx)
+	resp, err := request.Do(ctx)
 
 	if err != nil {
 		return err
@@ -87,6 +78,6 @@ func addMemberToGroup(ctx context.Context, client *client.APIClient, companyID, 
 		return err
 	}
 
-	fmt.Println("the users has been added to the group")
+	fmt.Printf("user %s successfully removed\n", userID)
 	return nil
 }